test(Properties) : use prod data

+++ src/test/java/com/myssteriion/blindtest/properties/RoundContentPropertiesTest.java
@@ -15,15 +15,15 @@ public class RoundContentPropertiesTest extends AbstractTest {
     @Test
     public void getterSetter() {
 
-        Assert.assertEquals( new Integer(5), props.getClassicNbMusics() );
+        Assert.assertEquals( new Integer(20), props.getClassicNbMusics() );
         Assert.assertEquals( new Integer(100), props.getClassicNbPointWon() );
 
-        Assert.assertEquals( new Integer(5), props.getChoiceNbMusics() );
+        Assert.assertEquals( new Integer(4), props.getChoiceNbMusics() );
         Assert.assertEquals( new Integer(100), props.getChoiceNbPointWon() );
         Assert.assertEquals( new Integer(50), props.getChoiceNbPointBonusWon() );
         Assert.assertEquals( new Integer(-50), props.getChoicenNPointMalusLoose() );
 
-        Assert.assertEquals( new Integer(5), props.getThiefNbMusics() );
+        Assert.assertEquals( new Integer(20), props.getThiefNbMusics() );
         Assert.assertEquals( new Integer(100), props.getThiefNbPointWon() );
         Assert.assertEquals( new Integer(-100), props.getThiefNbPointLoose() );
     }
